requirements/base.txt
Django>=1.7,<1.9

django-bootstrap-toolkit>=2.15.0
django-compressor>=1.5
  django-appconf>=1.0.1
django-selectable>=0.9.0

psycopg2>=2.6
python-dateutil>=2.4.2
pytz>=2015.2
six>=1.9.0
